backend/requirements/base.txt
Django>=4.2,<5.1
djangorestframework>=3.14,<3.16
django-cors-headers>=4.3,<5
djangorestframework-simplejwt>=5.3,<6
